another rename and readme edit
